# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sistema de Gestão Empresarial API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tags:
    get:
      operationId: getTags
      tags: [tags]
      summary: List tags
      responses:
        "200":
          description: Tags list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tag"
    post:
      operationId: createTag
      tags: [tags]
      summary: Create tag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTagRequest"
      responses:
        "201":
          description: Tag created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
  /tags/{id}:
    put:
      operationId: updateTag
      tags: [tags]
      summary: Update tag
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTagRequest"
      responses:
        "200":
          description: Tag updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
    delete:
      operationId: deleteTag
      tags: [tags]
      summary: Delete tag
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tag deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /tags/{id}/report:
    get:
      operationId: getTagReport
      tags: [tags]
      summary: Get sales and stock report by tag
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tag report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagReport"
  /sales:
    get:
      operationId: getSales
      tags: [sales]
      summary: List sales
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: customerId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Sales list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesPage"
    post:
      operationId: createSale
      tags: [sales]
      summary: Create sale
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSaleRequest"
      responses:
        "201":
          description: Sale created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
  /sales/{id}:
    get:
      operationId: getSale
      tags: [sales]
      summary: Get sale
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sale
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
    delete:
      operationId: deleteSale
      tags: [sales]
      summary: Delete sale
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sale deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /users:
    get:
      operationId: getUsers
      tags: [users]
      summary: List users
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsersPage"
    post:
      operationId: createUser
      tags: [users]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    put:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, manager, seller]
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, role, active, createdAt]
    CreateUserRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
        role:
          type: string
          enum: [admin, manager, seller]
      required: [name, email, password, role]
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
        role:
          type: string
          enum: [admin, manager, seller]
        active:
          type: boolean
    UsersPage:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [data, total, page, limit]
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        productCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, productCount, createdAt]
...
    Tag:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        color:
          type: string
        productCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, color, productCount, createdAt]
    CreateTagRequest:
      type: object
      properties:
        name:
          type: string
        color:
          type: string
      required: [name, color]
    TagReport:
      type: object
      properties:
        tag:
          $ref: "#/components/schemas/Tag"
        totalRevenue:
          type: number
        totalSales:
          type: integer
        totalStock:
          type: integer
        products:
          type: array
          items:
            $ref: "#/components/schemas/TagProductStats"
      required: [tag, totalRevenue, totalSales, totalStock, products]
    TagProductStats:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        stock:
          type: integer
        totalSales:
          type: integer
        totalRevenue:
          type: number
      required: [id, name, stock, totalSales, totalRevenue]
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
        costPrice:
          type: number
        stock:
          type: integer
        categoryId:
          type: integer
          nullable: true
        category:
          $ref: "#/components/schemas/Category"
          nullable: true
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, name, price, costPrice, stock, tags, active, createdAt]
...
    SaleItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
          nullable: true
        quantity:
          type: integer
        unitPrice:
          type: number
        subtotal:
          type: number
      required: [id, productId, quantity, unitPrice, subtotal]
    Sale:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
          nullable: true
        customer:
          $ref: "#/components/schemas/Customer"
          nullable: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItem"
        total:
          type: number
        discount:
          type: number
        paymentMethod:
          type: string
          enum: [cash, credit_card, debit_card, pix, transfer]
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, items, total, discount, paymentMethod, createdAt]
    CreateSaleRequest:
      type: object
      properties:
        customerId:
          type: integer
          nullable: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/CreateSaleItemRequest"
        discount:
          type: number
        paymentMethod:
          type: string
          enum: [cash, credit_card, debit_card, pix, transfer]
        notes:
          type: string
          nullable: true
      required: [items, paymentMethod]
    CreateSaleItemRequest:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
        unitPrice:
          type: number
      required: [productId, quantity, unitPrice]
    SalesPage:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Sale"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [data, total, page, limit]
    Customer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        document:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        totalPurchases:
          type: number
        purchaseCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, totalPurchases, purchaseCount, createdAt]
...